requirements.txt
numpy
Pillow
torch
torchvision
pyyaml
